unify RtpQueue as well, with tiny ifdefs
